block: queue work on power efficient wq
Block layer uses workqueues for multiple purposes. There is no real
dependency
of scheduling these on the cpu which scheduled them.

On a idle system, it is observed that and idle cpu wakes up many times
just to
service this work. It would be better if we can schedule it on a cpu
which the
scheduler believes to be the most appropriate one.

This patch replaces normal workqueues with power efficient versions.

Cc: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Tejun Heo <tj@kernel.org>
[zwliew: Backport to 3.4]
Signed-off-by: Zhao Wei Liew <zhaoweiliew@gmail.com>